+++ analysis_options.yaml
@@ -3,3 +3,9 @@ include: package:lint/analysis_options.yaml
 analyzer:
   errors:
     avoid_print: ignore
+    prefer_const_constructors: ignore
+    prefer_const_constructors_in_immutables: ignore
+    prefer_const_declarations: ignore
+    prefer_const_literals_to_create_immutables: ignore
+    sort_constructors_first: ignore
+    sort_child_properties_last: ignore
