+++ lib/UI/constants.dart
@@ -1,4 +1,4 @@
 import 'dart:ui';
 
 const kDarkBlue = Color(0xFF27407A);
-const klightBlue = Color(0xFF147FC3);
+const kLightBlue = Color(0xFF147FC3);
